// generated/prisma/client/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const DurationUnit = {
  DAYS: 'DAYS',
  MONTHS: 'MONTHS',
  YEARS: 'YEARS'
} as const

export type DurationUnit = (typeof DurationUnit)[keyof typeof DurationUnit]


export const RoleName = {
  SUPER_ADMIN: 'SUPER_ADMIN',
  ADMIN: 'ADMIN',
  USER: 'USER',
  CLIENT: 'CLIENT'
} as const

export type RoleName = (typeof RoleName)[keyof typeof RoleName]


export const TokenType = {
  EMAIL_VERIFY: 'EMAIL_VERIFY',
  PASSWORD_RESET: 'PASSWORD_RESET'
} as const

export type TokenType = (typeof TokenType)[keyof typeof TokenType]


export const Industry = {
  Education: 'Education',
  HealthCare: 'HealthCare'
} as const

export type Industry = (typeof Industry)[keyof typeof Industry]


export const UserStatus = {
  Active: 'Active',
  InActive: 'InActive',
  Invited: 'Invited',
  Deactivate: 'Deactivate',
  Suspend: 'Suspend'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const OrganizationStatus = {
  Active: 'Active',
  InActive: 'InActive',
  Pending: 'Pending',
  Suspend: 'Suspend',
  Archive: 'Archive'
} as const

export type OrganizationStatus = (typeof OrganizationStatus)[keyof typeof OrganizationStatus]


export const CreditType = {
  Credit: 'Credit',
  Debit: 'Debit'
} as const

export type CreditType = (typeof CreditType)[keyof typeof CreditType]


export const SubscriptionType = {
  Monthly: 'Monthly',
  Annual: 'Annual'
} as const

export type SubscriptionType = (typeof SubscriptionType)[keyof typeof SubscriptionType]


export const InvoiceStatus = {
  PENDING: 'PENDING',
  PAID: 'PAID',
  CANCELLED: 'CANCELLED'
} as const

export type InvoiceStatus = (typeof InvoiceStatus)[keyof typeof InvoiceStatus]


export const PaymentType = {
  CASH: 'CASH',
  BANK_TRANSFER: 'BANK_TRANSFER',
  CHEQUE: 'CHEQUE',
  CREDIT_CARD: 'CREDIT_CARD',
  DEBIT_CARD: 'DEBIT_CARD',
  UPI: 'UPI',
  OTHER: 'OTHER'
} as const

export type PaymentType = (typeof PaymentType)[keyof typeof PaymentType]


export const CreditExpiry = {
  OneMonth: 'OneMonth',
  ThreeMonths: 'ThreeMonths',
  SixMonths: 'SixMonths',
  TwelveMonths: 'TwelveMonths',
  EighteenMonths: 'EighteenMonths'
} as const

export type CreditExpiry = (typeof CreditExpiry)[keyof typeof CreditExpiry]


export const Gender = {
  MALE: 'MALE',
  FEMALE: 'FEMALE',
  OTHER: 'OTHER'
} as const

export type Gender = (typeof Gender)[keyof typeof Gender]


export const MembershipStatus = {
  INVITED: 'INVITED',
  ACTIVE: 'ACTIVE',
  SUSPENDED: 'SUSPENDED',
  REMOVED: 'REMOVED'
} as const

export type MembershipStatus = (typeof MembershipStatus)[keyof typeof MembershipStatus]